Web/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
requests
pyttsx3
moviepy
Pillow
numpy